backend/api_server/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
bcrypt>=4.1.0
pyjwt>=2.8.0
